isis: extract RIB / SPF-orchestration into rib.rs
Final move in the four-PR split of the oversized isis/inst.rs.
Move everything that translates an SPF result into FIB updates
into a new rib.rs:

  - spf_timer, spf_schedule (debounced SPF kickoff)
  - SpfRoute, SpfNexthop, SpfRouteV6, SpfNexthopV6, SpfIlm
  - BACKUP_METRIC_OFFSET; DiffResult{,V6}, DiffIlmResult type aliases
  - nhop_to_nexthop_uni{,_v6}, backup_to_nexthop_uni{,_v6}
  - make_rib_entry{,_v6}, build_rib_nexthop, make_ilm_entry
  - diff_apply, diff_apply_v6, diff_ilm_apply
  - build_adjacency_ilm
  - build_rib_from_spf, build_rib_from_spf_v6, ipv6_capable_set
  - apply_routing_updates, perform_spf_calculation, mpls_route
  - the remaining build_rib_nexthop / make_rib_entry / make_rib_entry_v6
    unit tests

Message enum + Display stay in inst.rs — it's the IS-IS instance's
event channel, not RIB code. inst.rs keeps the Isis struct, IsisTop,
event_loop, config plumbing, and the SR subscription reconciler.

Consumers (show.rs, nfsm.rs, packet.rs, lsdb.rs, inst.rs itself) now
pull spf_schedule / SpfRoute / SpfNexthop / etc. from super::rib
instead of super::inst.

inst.rs drops 2030 → 863 lines. Cumulative across the four PRs:
3525 → 863 (~75% reduction) with tilfa.rs (341) + graph.rs (474) +
lsp.rs (823) + rib.rs (1185) carrying the extracted code.
