+++ zebra-rs/src/isis/mod.rs
@@ -52,3 +52,5 @@ pub use flood::LspFlood;
 pub mod tilfa;
 
 pub mod graph;
+
+pub mod rib;
