+++ zebra-rs/src/isis/show.rs
@@ -6,7 +6,7 @@ use isis_packet::{IsisProto, IsisSysId, IsisTlv, Nsap};
 use prefix_trie::PrefixMap;
 use serde::Serialize;
 
-use super::inst::{SpfNexthop, SpfNexthopV6, SpfRoute, SpfRouteV6};
+use super::rib::{SpfNexthop, SpfNexthopV6, SpfRoute, SpfRouteV6};
 use super::tilfa::{RepairPathMpls, RepairPathSrv6};
 use super::{Isis, inst::ShowCallback};
